Refactor cave node data to use a common interface for rooms and connections.
