--- Assets/Scripts/CaveV2/CaveGraph/CaveNodeData.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: c3aaba5e797445389ad687fb92c6d020
-timeCreated: 1655329037
