+++ Assets/Scripts/CaveV2/CaveGraph/CaveNodeDataDebugComponent.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using BML.Scripts.CaveV2.CaveGraph.NodeData;
 using Shapes;
 using Sirenix.OdinInspector;
 using UnityEngine;
