+++ Assets/Scripts/CaveV2/CaveGraph/NodeData.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d082076108ef4fb4ae36704a130fab7c
+timeCreated: 1669095846
