+++ Assets/Scripts/CaveV2/CaveGraph/NodeData/CaveNodeConnectionData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b2626fbc3bb647e99f97ab37052241dc
+timeCreated: 1669096049
