+++ Assets/Scripts/CaveV2/CaveGraph/NodeData/CaveNodeData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: cb6f423ef040411787b3b9cd506cdf7b
+timeCreated: 1669096196
