+++ Assets/Scripts/CaveV2/CaveGraph/NodeData/ICaveNodeData.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 0d90bf2505d545ed80af08c2fde4f336
+timeCreated: 1669095974
