+++ Assets/Scripts/CaveV2/SpawnObjects/LevelObjectSpawner.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using BML.ScriptableObjectCore.Scripts.Events;
 using BML.Scripts.CaveV2.CaveGraph;
+using BML.Scripts.CaveV2.CaveGraph.NodeData;
 using BML.Scripts.CaveV2.MudBun;
 using BML.Scripts.Utils;
 using KinematicCharacterController;
